Moving Maass form files around
