+++ lmfdb/maass_forms/web_maassform.py
@@ -12,11 +12,16 @@ def character_link(level, conrey_index):
 
 
 def fricke_pretty(fricke_eigenvalue):
-    return "$%+d$" % fricke_eigenvalue if fricke_eigenvalue else ""
+    if fricke_eigenvalue == 0:
+        return "not computed rigorously"
+    elif fricke_eigenvalue == 1:
+        return "$+1$"
+    else:
+        return "$-1$"
 
 
 def symmetry_pretty(symmetry):
-    return "even" if symmetry == 1 else ("odd" if symmetry == -1 else "")
+    return "even" if symmetry == 0 else ("odd" if symmetry == 1 else "")
 
 
 def th_wrap(kwl, title):
@@ -35,68 +40,117 @@ def td_wrapr(val):
     return '    <td align="right">%s</td>' % val
 
 
+def coeff_error_notation(coeff, error):
+    r"""Web coefficient and error display, with trunctation"""
+    if error == -1:
+        return r"\mathrm{unknown}"
+    coeffpart = "%+.8f" % coeff
+    if error == 0:
+        errorpart = ""
+    else:
+        base, negexponent = mantissa_and_exponent(error)
+        negexponent = min(negexponent, 8)
+        errorpart = r" \pm " + exponential_form(base, negexponent, digits_to_show=3)
+    return coeffpart + errorpart
+
+
+def mantissa_and_exponent(number):
+    """
+    Returns (mantissa, exponent) where number = mantissa * 10^(-exponent).
+
+    The input should be a RealLiteral.
+    """
+    if number > 1:
+        return (number, 0)
+    snum = str(number)
+    if '.' not in snum:
+        return (number, 0)
+    pre, post = snum.split('.')
+    # determine the exponent
+    idx = 0
+    for idx, c in enumerate(post):
+        if c != '0':
+            break
+    negative_exponent = idx + 1
+    mantissa = post[idx] + "." + post[idx+1:]
+    return (mantissa, negative_exponent)
+
+
+def exponential_form(mantissa, negative_exponent, digits_to_show=9):
+    """
+    Format the number `mantissa * 10^(-negative_exponent)` nicely.
+    """
+    if negative_exponent == 0:
+        return mantissa[:digits_to_show]
+    return mantissa[:digits_to_show] + r" \cdot 10^{-" + str(negative_exponent) + "}"
+
+
 def parity_text(val):
     return 'odd' if val == -1 else 'even'
 
 
-class WebMaassForm():
+def short_label(label):
+    # We shorten some labels from 5 components to 2 for simplicity (all of the Maass forms in the initial draft of the database had the same values for the second, third and fifth part of the label)
+    pieces = label.split(".")
+    if len(pieces) == 5 and pieces[1] == '0' and pieces[2] == '1' and pieces[4] == '1':
+        return f"{pieces[0]}.{pieces[3]}"
+    return label
+
+
+def long_label(label):
+    # Undo the transformation from short_label
+    pieces = label.split(".")
+    if len(pieces) == 2:
+        return f"{pieces[0]}.0.1.{pieces[1]}.1"
+    return label
+
+
+class WebRigorMaassForm():
     def __init__(self, data):
         self.__dict__.update(data)
         self._data = data
-        self.portrait = db.maass_portraits.lookup(self.label, projection="portrait")
+        # TODO figure out how to handle portraits appropriately
+        #self.portrait = db.maass_portraits.lookup(self.label, projection="portrait")
 
     @staticmethod
     def by_label(label):
-        data = db.maass_newforms.lookup(label)
+        data = db.maass_rigor.lookup(label)
         if data is None:
             raise KeyError("Maass newform %s not found in database." % (label))
-        return WebMaassForm(data)
-
-    @staticmethod
-    def by_maass_id(maass_id):
-        return WebMaassForm.by_label(maass_id)
+        return WebRigorMaassForm(data)
 
     @property
     def label(self):
-        return self.maass_id  # TODO: we should revisit this at some point
-
-    # next_maass_form and prev_maass_form below are provided for backward compatibility
-    # they are referenced in lfunctions but are not currently used
-
-    @property
-    def next_maass_form(self):
-        # return the "next" maass form of the save level, character, and weight with spectral parameter approximately >= our spectral _parameter
-        # forms with the same spectral parameter are sorted by maass_id
-        query = {'level': self.level, 'weight': self.weight, 'conrey_index': self.conrey_index, 'spectral_parameter': self.spectral_parameter, 'maass_id': {'$gt': self.maass_id}}
-        forms = db.maass_newforms.esearch(query, sort=["maass_id"], projection='maass_id', limit=1)
-        if forms:
-            return forms[0]
-        query = {'level': self.level, 'weight': self.weight, 'conrey_index': self.conrey_index, 'spectral_parameter': {'$gt': self.spectral_parameter}}
-        forms = db.maass_forms.search(query, sort=["spectral_parameter", "maass_id"], projection='maass_id', limit=1)
-        return forms[0] if forms else None
-
-    @property
-    def prev_maass_form(self):
-        query = {'level': self.level, 'weight': self.weight, 'conrey_index': self.conrey_index, 'spectral_parameter': self.spectral_parameter, 'maass_id': {'$lt': self.maass_id}}
-        forms = db.maass_newforms.esearch(query, sort=["maass_id"], projection='maass_id', limit=1)
-        if forms:
-            return forms[0]
-        query = {'level': self.level, 'weight': self.weight, 'conrey_index': self.conrey_index, 'spectral_parameter': {'$lt': self.spectral_parameter}}
-        forms = db.maass_forms.search(query, sort=["spectral_parameter", "maass_id"], projection='maass_id', limit=1)
-        return forms[0] if forms else None
+        return self.maass_label
 
     @property
     def coeffs(self):
         return [RR(c) for c in self.coefficients]
 
     @property
+    def spectral_index(self):
+        return self.nspec
+
+    @property
+    def web_spectral_error(self):
+        str_error = "%E" % self.spectral_error
+        base, _, exponent = str_error.partition("E")
+        return rf"{base} \cdot 10^{{{exponent}}}"
+
+    @property
     def title(self):
-        return r"Maass form on \(\Gamma_0(%d)\) with \(R=%s\)" % (self.level, self.spectral_parameter)
+        digits_to_show = 10
+        return (rf"Maass form {short_label(self.label)} on \(\Gamma_0({self.level})\) "
+                rf"with \(R={str(self.spectral_parameter)[:digits_to_show]}\)")
 
     @property
     def properties(self):
-        props = [(None, '<img src="{0}" width="200" height="150" style="margin:10px;"/>'.format(self.portrait))] if self.portrait is not None else []
-        props += [('Level', prop_int_pretty(self.level)),
+        # props = [
+        #   (None, '<img src="{0}" width="200" height="150" style="margin:10px;"/>'.format(self.portrait))
+        # ] if self.portrait is not None else []
+        props = []
+        props += [('Label', short_label(self.label)),
+                  ('Level', prop_int_pretty(self.level)),
                   ('Weight', prop_int_pretty(self.weight)),
                   ('Character', self.character_label),
                   ('Symmetry', self.symmetry_pretty),
@@ -143,11 +197,17 @@ class WebMaassForm():
 
     @property
     def friends(self):
-        return [("L-function", "/L" + url_for(".by_label", label=self.label))]
+        friendlist = []
+        if self.nspec > 1:
+            prevlabel = f"{self.level}.{self.weight}.{self.conrey_index}.{self.nspec - 1}.1"
+            friendlist += [("Previous Maass form", url_for(".by_label", label=prevlabel))]
+        if self.nspec < db.maass_rigor.count(query={'level':self.level}):
+            nextlabel = f"{self.level}.{self.weight}.{self.conrey_index}.{self.nspec + 1}.1"
+            friendlist += [("Next Maass form", url_for(".by_label", label=nextlabel))]
+        friendlist += [("L-function not computed", '')]
+        return friendlist
 
-    def coefficient_table(self, rows=20, cols=5):
-        if not self.coefficients:
-            return ""
+    def coefficient_table(self, rows=20, cols=3):
         n = len(self.coefficients)
         assert rows > 0 and cols > 0
         table = ['<table class="ntdata"><thead><tr><th></th></tr></thead><tbody>']
@@ -158,8 +218,13 @@ class WebMaassForm():
             for j in range(cols):
                 if i * cols + j >= n:
                     break
-                table.append(td_wrapr(r"\(a_{%d}=%+.9f\)" % (i * cols + j + 1,
-                                                             self.coefficients[i * cols + j])))
+                table.append(
+                    td_wrapl(r"\(a_{%d}= %s \)"
+                             % (i * cols + j + 1,
+                                coeff_error_notation(
+                                    self.coefficients[i * cols + j],
+                                    self.coefficient_errors[i * cols + j]
+                                ))))
             table.append('</tr>')
         table.append('</tbody></table>')
         if rows * cols < n:
@@ -170,30 +235,32 @@ class WebMaassForm():
 
 
 class MaassFormDownloader(Downloader):
-    table = db.maass_newforms
-    title = 'Maass forms'
+    title = 'Rigorous Maass forms'
 
     def download(self, label, lang='text'):
-        data = db.maass_newforms.lookup(label)
+        table = db.maass_rigor
+        data = table.lookup(label)
         if data is None:
             return abort(404, "Maass form %s not found in the database" % label)
-        for col in db.maass_newforms.col_type:
-            if db.maass_newforms.col_type[col] == "numeric" and data.get(col):
+        for col in table.col_type:
+            if table.col_type[col] == "numeric" and data.get(col):
                 data[col] = str(data[col])
-            if db.maass_newforms.col_type[col] == "numeric[]" and data.get(col):
+            if table.col_type[col] == "numeric[]" and data.get(col):
                 data[col] = [str(data[col][n]) for n in range(len(data[col]))]
         return self._wrap(Json.dumps(data),
-                          label,
+                          "maass." + label,
                           lang=lang,
-                          title='All stored data for Maass form %s,' % (label))
+                          title='All stored data for Rigorous Maass form %s,' % (label))
 
     def download_coefficients(self, label, lang='text'):
-        data = db.maass_newforms.lookup(label, projection="coefficients")
+        table = db.maass_rigor
+        data = table.lookup(label, projection=["coefficients", "coefficient_errors"])
         if data is None:
-            return abort(404, "Coefficient data for Maass form %s not found in the database" % label)
-        c = data
-        data = [str(c[n]) for n in range(len(c))]
-        return self._wrap(Json.dumps(data).replace('"', ''),
-                          label + '.coefficients',
+            return abort(404, "Coefficient data for Rigorous Maass form %s not found in the database" % label)
+        coeffs = data["coefficients"]
+        errors = data["coefficient_errors"]
+        retdata = [str(coeffs[n]) + " +- " + str(errors[n]) for n in range(len(coeffs))]
+        return self._wrap(Json.dumps(retdata).replace('"', ''),
+                          "maass." + label + '.coefficients',
                           lang=lang,
-                          title='Coefficients for Maass form %s,' % (label))
+                          title='Coefficients for Rigorous Maass form %s,' % (label))
